requirements.txt
fastapi==0.98.0
uvicorn
asyncio
aiohttp==3.8.5

pandas==2.0.3

gunicorn
uvloop
httptools

python-jose
passlib
python-multipart
bcrypt

jinja2
aiofiles

pydantic==1.10.4
python-dotenv
google
google-cloud-firestore
python-dotenv==1.0.0
google-cloud-logging==3.4.0

